feat: chỉnh sửa lại màn profile
