+++ app/src/main/java/com/pwhs/quickmem/presentation/app/library/study_set/ListStudySetScreen.kt
@@ -43,7 +43,7 @@ import com.pwhs.quickmem.domain.model.color.ColorModel
 import com.pwhs.quickmem.domain.model.study_set.GetStudySetResponseModel
 import com.pwhs.quickmem.domain.model.subject.SubjectModel
 import com.pwhs.quickmem.presentation.ads.BannerAds
-import com.pwhs.quickmem.presentation.app.profile.profile.ProfileViewModel
+import com.pwhs.quickmem.presentation.app.profile.ProfileViewModel
 import com.pwhs.quickmem.util.toColor
 
 @OptIn(ExperimentalMaterial3Api::class)
