--- app/src/main/java/com/pwhs/quickmem/presentation/app/profile/ProfileAndStaticsScreen.kt
@@ -1,101 +1,0 @@
-package com.pwhs.quickmem.presentation.app.profile
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.MaterialTheme.colorScheme
-import androidx.compose.material3.MaterialTheme.typography
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Tab
-import androidx.compose.material3.TabRow
-import androidx.compose.material3.TabRowDefaults.SecondaryIndicator
-import androidx.compose.material3.TabRowDefaults.tabIndicatorOffset
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.hilt.navigation.compose.hiltViewModel
-import com.pwhs.quickmem.presentation.app.profile.profile.TabProfileScreen
-import com.pwhs.quickmem.presentation.app.profile.statics.TabStatisticsScreen
-import com.ramcosta.composedestinations.annotation.Destination
-import com.ramcosta.composedestinations.annotation.RootGraph
-import com.ramcosta.composedestinations.navigation.DestinationsNavigator
-
-
-@Composable
-@Destination<RootGraph>
-fun ProfileScreen(
-    modifier: Modifier = Modifier,
-    viewModel: ProfileAndStaticsViewModel = hiltViewModel(),
-    navigator: DestinationsNavigator
-) {
-
-    val uiState by viewModel.uiState.collectAsState()
-
-    LaunchedEffect(key1 = true) {
-        viewModel.uiEvent.collect { event ->
-            when (event) {
-                is ProfileAndStaticsUiEventUiEvent.LoadProfile -> {}
-            }
-        }
-    }
-
-    var tabIndex by remember { mutableIntStateOf(0) }
-    val tabTitles = listOf("Profile", "Statistics")
-
-    Scaffold(
-        topBar = {
-        }
-    ) { innerPadding ->
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(Color.Transparent)
-                .padding(innerPadding)
-        ) {
-            TabRow(
-                selectedTabIndex = tabIndex,
-                indicator = { tabPositions ->
-                    SecondaryIndicator(
-                        Modifier.tabIndicatorOffset(tabPositions[tabIndex]),
-                        color = colorScheme.primary
-                    )
-                }
-            ) {
-                tabTitles.forEachIndexed { index, title ->
-                    Tab(
-                        text = {
-                            Text(
-                                title,
-                                style = typography.titleMedium.copy(
-                                    fontWeight = FontWeight.Bold,
-                                    color = if (tabIndex == index) Color.Black else Color.Gray
-                                )
-                            )
-                        },
-                        selected = tabIndex == index,
-                        onClick = { tabIndex = index }
-                    )
-                }
-            }
-
-            when (tabIndex) {
-                0 -> TabProfileScreen(navigator = navigator)
-                1 -> TabStatisticsScreen()
-            }
-        }
-    }
-}
-
-
-
-
-
