--- app/src/main/java/com/pwhs/quickmem/presentation/app/profile/ProfileAndStaticsUiAction.kt
@@ -1,5 +1,0 @@
-package com.pwhs.quickmem.presentation.app.profile
-
-sealed class ProfileAndStaticsUiAction {
-    data object LoadProfile : ProfileAndStaticsUiAction()
-}
