--- app/src/main/java/com/pwhs/quickmem/presentation/app/profile/ProfileAndStaticsUiEvent.kt
@@ -1,5 +1,0 @@
-package com.pwhs.quickmem.presentation.app.profile
-
-sealed class ProfileAndStaticsUiEventUiEvent {
-    data object LoadProfile : ProfileAndStaticsUiEventUiEvent()
-}
