--- app/src/main/java/com/pwhs/quickmem/presentation/app/profile/ProfileAndStaticsUiState.kt
@@ -1,7 +1,0 @@
-package com.pwhs.quickmem.presentation.app.profile
-
-data class ProfileAndStaticsUiState(
-    val profileData: String = "",
-    val isLoading: Boolean = false,
-    val errorMessage: String? = null
-)
