+++ app/src/main/java/com/pwhs/quickmem/presentation/app/profile/ProfileScreen.kt
@@ -1,0 +1,89 @@
+package com.pwhs.quickmem.presentation.app.profile
+
+import android.net.Uri
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
+import com.pwhs.quickmem.presentation.app.profile.component.ProfileTab
+import com.pwhs.quickmem.presentation.app.profile.component.ProfileTopBar
+import com.ramcosta.composedestinations.annotation.Destination
+import com.ramcosta.composedestinations.annotation.RootGraph
+import com.ramcosta.composedestinations.navigation.DestinationsNavigator
+
+
+@Composable
+@Destination<RootGraph>
+fun ProfileScreen(
+    modifier: Modifier = Modifier,
+    viewModel: ProfileViewModel = hiltViewModel(),
+    navigator: DestinationsNavigator,
+
+    ) {
+    val name by viewModel.nameState.collectAsState(initial = "")
+    val avatarUrl by viewModel.avatarUrlState.collectAsState(initial = "")
+    val context = LocalContext.current
+
+    val launcher = rememberLauncherForActivityResult(
+        contract = ActivityResultContracts.GetContent()
+    ) { uri: Uri? ->
+        uri?.let {
+            viewModel.updateAvatar(it.toString())
+        }
+    }
+
+    Scaffold(
+        topBar = {
+            ProfileTopBar(
+                onUpgradeClick = {
+            })
+        },
+        content = { paddingValues ->
+            Column(
+                modifier = modifier
+                    .fillMaxSize()
+                    .padding(paddingValues)
+                    .verticalScroll(rememberScrollState())
+                    .padding(horizontal = 16.dp)
+                    .background(MaterialTheme.colorScheme.background),
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.Top
+            ) {
+                Spacer(modifier = Modifier.height(26.dp))
+                ProfileTab(
+                    name = name,
+                    avatarUrl = avatarUrl,
+                    navigator = navigator,
+                    onImageClick = {
+                        launcher.launch("image/*")
+                    }
+                )
+            }
+        }
+    )
+}
+
+
+
+
+
+
+
+
