+++ app/src/main/java/com/pwhs/quickmem/presentation/app/profile/ProfileUiAction.kt
@@ -1,0 +1,5 @@
+package com.pwhs.quickmem.presentation.app.profile
+
+sealed class ProfileUiAction {
+    data object LoadProfile : ProfileUiAction()
+}
