+++ app/src/main/java/com/pwhs/quickmem/presentation/app/profile/component/ProfileTab.kt
@@ -1,0 +1,74 @@
+package com.pwhs.quickmem.presentation.app.profile.component
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import coil.compose.AsyncImage
+import coil.request.ImageRequest
+import com.pwhs.quickmem.R
+import com.ramcosta.composedestinations.annotation.Destination
+import com.ramcosta.composedestinations.annotation.RootGraph
+import com.ramcosta.composedestinations.generated.destinations.SettingsScreenDestination
+import com.ramcosta.composedestinations.navigation.DestinationsNavigator
+
+@Destination<RootGraph>
+@Composable
+fun ProfileTab(
+    name: String,
+    avatarUrl: String,
+    navigator: DestinationsNavigator,
+    onImageClick: () -> Unit
+) {
+    Column(
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.spacedBy(16.dp),
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        AsyncImage(
+            model = ImageRequest.Builder(LocalContext.current)
+                .data(avatarUrl.ifEmpty { null })
+                .placeholder(R.drawable.default_avatar)
+                .error(R.drawable.default_avatar)
+                .build(),
+            contentDescription = "User Avatar",
+            modifier = Modifier
+                .size(80.dp)
+                .clip(CircleShape)
+                .background(Color.Gray)
+                .clickable { onImageClick() },
+            contentScale = ContentScale.Crop
+        )
+
+        Text(
+            text = name,
+            fontSize = 20.sp,
+            fontWeight = FontWeight.Bold
+        )
+
+        Spacer(modifier = Modifier.height(16.dp))
+
+        SettingsButton(
+            text = "Cài đặt của bạn",
+            onClick = { navigator.navigate(SettingsScreenDestination) }
+        )
+
+        Spacer(modifier = Modifier.height(32.dp))
+    }
+}
