--- app/src/main/java/com/pwhs/quickmem/presentation/app/profile/statics/TabStaticsScreen.kt
@@ -1,24 +1,0 @@
-package com.pwhs.quickmem.presentation.app.profile.statics
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-
-@Composable
-fun TabStatisticsScreen(
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Text("Statistics Screen")
-        // Add other content related to statistics here
-    }
-}
