+++ app/src/main/java/com/pwhs/quickmem/presentation/app/settings/SettingsViewModel.kt
@@ -1,0 +1,22 @@
+package com.pwhs.quickmem.presentation.app.settings
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.pwhs.quickmem.core.datastore.AppManager
+import com.pwhs.quickmem.core.datastore.TokenManager
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+@HiltViewModel
+class SettingsViewModel @Inject constructor(
+    private val appManager: AppManager,
+    private val tokenManager: TokenManager
+) : ViewModel() {
+
+    fun logout() {
+        viewModelScope.launch {
+            tokenManager.clearTokens()
+        }
+    }
+}
